Update patch set 2
Patch Set 2:

Thank you for the update.  To ensure the issue numbers are found by the Gerrit/Jira integration please add a line with exactly this format:

Issue: CD-2015, CD-2029

One significant problem is that these issues are not in the Linux Foundation Jira, but we'll get to that.

Patch-set: 2