Update patch set 4
Patch Set 4:

(1 comment)

Header and footer can be separated from the email body templates. I am waiting for the actual template which I can use.
For #4 we are using the content type as "text/html". Email Client should be able to convert that to plain text.

For CLI or other components we have exposed a rest end point for now which can be used to trigger notification. Only templates needs to be generated for other kind of emails/components

Patch-set: 4
Label: Code-Review=0