Update patch set 5
Patch Set 5:

> I am fine with moving the notification functionality entirely to a
 > new component/docker ONLY IF the portal removes its ability to send
 > email AND uses that new notification component when it needs to
 > send email.

That was the initial thinking when we started the implementation of notification of intermediate steps for onboarding and validation but unfortunately we ended up in creating a schema in CDS which is not generic enough to hold all kinds of notifications from different component.
But if we can have a separate component to handle the notification then portal will use same component to send any kind of notification. There is no sense in keeping same login in two different components.

Patch-set: 5
Reviewer: Gerrit User 56 <56@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0