Disable 'Add to Catalog' when on-boarding
Issue-ID: ACUMOS-742
Description: Disables the add to catalog button whenever on-boarding is
started(so once you click it). It's re-enabled once it returns a step failed or finishes
on-boarding. It also properly scrolls to the top where the message shows
up once you click on add to catalog. Extended the time for a few seconds
so you can actually read it.

Change-Id: Ie83f3884c9e1821ab189527bec5d678c60d90efa
Signed-off-by: Raphael Sarmiento <RS0C74398@techmahindra.com>