Add validation for admin supporting content
Issue-ID: ACUMOS-913
Added dynamic "characters left" span below supporting content in
creation of new slides for admin content. If 140-character limit is
exceeded, text is changed to red and a warning is appended to the span.
"Done" button is also disabled until character limit is reduced, as
extra precaution. May be expandable later to cover other validation
checks.

Change-Id: I48f6c101a0382990538bf92a72ea658c6ef569aa
Signed-off-by: swood8 <SW0C74397@TechMahindra.com>