Reduce user session duration for regular timeouts
Issue-ID: ACUMOS-966
Original issue states that session timeouts do not occur; code reveals
that timeouts do occur but are set to a week in duration. Duration
changed to configurable parameter in application.properties, defaulting
to 30 minutes, such that timeout can be set using the docker compose in
order to best address the security risks concerning persisting tokens.

Change-Id: Id1af0ee77308638d2305a4bc8d037ab59d820214
Signed-off-by: swood8 <SW0C74397@TechMahindra.com>