Echo model name after entry during onboarding
Issue-ID: ACUMOS-1200
Corrections made to onboarding page such that, when the user enters a
name for the model they wish to upload, the model name will be echoed
back to the user on the page. Minor style changes were also made to
the text for consistency with the echo of the toolkit type selected.

Change-Id: I23d93cc30017abc58452215c1112f9488a943d42
Signed-off-by: swood8 <SW0C74397@TechMahindra.com>