Fix where model documents lead to "Page Not Found"
Issue-ID: ACUMOS-1432
Fixed issue in code where improper use of ng-if was previously leading
to Page Not Found errors when trying to download Company docs from an
unpublished model. Unification of URLs used for fetching documents in
another change may have already fixed this issue, in which case this
change acts to clean up unnecessary/depreciated code related to the
issue.

Change-Id: Id994b14d5974d44314c55c37765c3db94c95db19
Signed-off-by: swood8 <SW0C74397@TechMahindra.com>