Update patch set 1
Patch Set 1:

> (12 comments)
 > 
 > Hope this helps

Commit Message, line 9: This has been the format for commit messages since I joined the project, and I have been asked to adhere to it. If the format has recently changed, it has not come to my attention as of yet. If you have an issue with the current format, please raise this concern with Mukesh.

AdminService.java, line 56: Could you better explain what it is that you are requesting in terms of Javadoc comments? I see no special actions I need to take within this class, and though I might be wrong, does the pom.xml not generate the Javadocs for us? As for why the indentation is inconsistent, I cannot say; it appears correct in Spring Tools.

AdminServiceImplTest.java, line 284: I included test cases for both the controller and service to stick to the conventions of the entire test suite; this should also help with keeping SonarQube percentages from dipping.

 > (12 comments)
 > 
 > Hope this helps

Commit Message, line 9: This has been the format for commit messages since I joined the project, and I have been instructed to adhere to it. If you have an issue with the current format, please notify Mukesh to see about having it changed, if it can be modified at all.

AdminService.java, line 56: I am not sure what it is you are asking for here. I see no special work done for other signatures in this file to specify Javadoc information, and furthermore, unless I am mistaken, I believe that the pom.xml automatically generates the Javadocs for us. Also, I am stumped by the inconsistent indentation as well; it looks normal in STS.

AdminServiceImpl.java, line 155: I feel that this conflicts your feelings from the JIRA thread, but okay, I can submit an improvement story ticket for a CDS call; instead of checking for the count on a single peer alone, however, perhaps it would be more advantageous for a call that will check and return counts for all peers in a single SQL statement, such that we do not have to continuously query the CDS. This shouldn't be a problem, as I believe the only other time we would need a count for subs on a single peer is when we actually want the full list of subscriptions for display.

AdminServiceTestImpl.java, line 284: My reason for including this test case was to adhere to the standards of the test suite, which appear to include tests for each method so as to keep SonarQube happy.

All other issues have been or will be addressed and applied in the next patch set.

Patch-set: 1