Update patch set 4
Patch Set 4:

> (2 comments)

I have talked with Mukesh, and he insists that the Issue ID belongs at the top of the body. I will be ignoring this comment on this and any other changes based on his feedback. If you have an issue, talk with him about this.

Concerning the regex, which is based on RFC 1123 hostname standards, it can identify a valid hostname and can accept an IP address as well. It cannot verify that a hostname can be resolved to an IP address, however; what you are asking for would require setting up a DNS service, which would require its own server space that, from previous talks with Mukesh, is not something we should do. And while there do exist a few possible ways to work around this, as far as I am aware after previous research, none of them are available to us within the scope of Angular.

Patch-set: 4