Update patch set 4
Patch Set 4:

I think the back-end should verify the host name, whether it can be resolved to an IP address, because that would be a valuable service to provide to the user via a Verify button.  I wasn't trying to ask for the front-end java script to do that.  It's entirely likely that the user's browser and the back-end machines have very different views of the network world.

Patch-set: 4