Update patch set 4
Patch Set 4:

> (3 comments)
 > 
 > Minor suggestions for improving maintainability

Revisiting AsyncServiceImpl, I could not use arrays in conjunction with switch cases, as it requires constants for case evaluations. Using your suggestion for an incremented list, however, I've decided to recode this with a "circular list" approach. The step name will now be compared against an array list of the step name strings; if it matches a step inside, we use the name of the next step in the circle and use the first step name if we were given the last step name, and if the step name we were given isn't in the list, it should default to the first step ("CreateSolution"). I will be testing this approach locally and have it pushed before end of day.

As for the constant in the test cases, that is my fault. I did create the constant to replace all uses of the string, I must have simply, in my haste, missed a single replacement. It will be fixed in the next patch set.

Patch-set: 4
Reviewer: Gerrit User 101 <101@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0