Update patch set 7
Patch Set 7: Code-Review-1

Came up with an alternative approach to the addition of a ServiceExceptions interface by instead adding one or two new PortalUtils methods that will take in the env property name, check if it is null and/or empty, throw the preformatted exception if it fails the check, and otherwise returns the String value if it passes. Waiting for word from Mukesh before I either finalize this change (which I now realize is missing the license for the new class) or otherwise switch to using the proposed setup above.

Patch-set: 7
Reviewer: Gerrit User 101 <101@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1