Update patch set 1
Patch Set 1:

> I see many front-end changes.  But what about back-end changes? 
 > You cannot rely on security thru obscurity.

We are passing only solutionId to get solution details (in MarketPlaceCatalogServiceController.java -> getSolutionsDetails() method). To fix this issue in backend (portal-be) we also need to pass revisionID to get solution details. I think CDS needs to change the existing API (dataServiceRestClient.getSolution(solutionId)) or create new API to get soltion details which takes revisionId as input. In url string we are passing solutionId and revisionId to see the solution details. But backend is using only solutionId to show solution details. 

Solution details example url is below. 

http://localhost:8085/#/marketSolutions?solutionId=3f3d1b9a-8b40-4bb8-af74-4e1953e0ab65&revisionId=60510a2c-3ec0-49b5-996e-9fa6a674d817#md-model-detail-template

Patch-set: 1