Update patch set 1
Patch Set 1:

> (23 comments)
 > 
 > Wish I could help you more with the front-end JS

I have made corrections based on your suggestions and will update with a new patch shortly after testing, but here are some comments I have on some of your suggested edits.

Concerning capital S's in the URLs, that was a carry-over from the original API used with the CMS; in wanting to keep things similar to the old API, I tried to keep some elements similar, including old URLs. Per your request, however, I have changed the letters to lowercase.

In admin.component.js, there is only one instance of the "data:image/jpeg;base64," string, so making it into a constant would be nice but perhaps not as efficient and helpful as it could be. And there is no mention of PNGs in that one line of code because its purpose is to represent "jpg" extensions with a valid "image/jpeg" mime type; PNGs will be given the correct mime type if a file possesses that extension instead.

In footer.component.js, I understand your concerns given the anti pattern. My concern is where we literally store the information as a JSON string within the MLPSiteContent object. There could be better ways to return this data from the backend, I will acknowledge that, but until I have an opportunity to discuss better returns from the backend with Mukesh, this format will stand for now on this first iteration of the new controller and service.

Patch-set: 1
Label: Code-Review=0