Update patch set 3
Patch Set 3:

> (17 comments)
 > 
 > My concerns are
 > 1. Is this change set solely focused on global site content like
 > carousel, contact, footer, terms?  It seems to touch
 > solution-specific stuff.
 > 2. Will the new method of fetching images allow the browser to
 > cache them, for faster page load?

After a bit of work, I have redesigned the controller such that we can use URLs that return the raw binary image, though unless there is additional work in the form of additional endpoints and perhaps a preliminary call for each image, we will always assume the returned image is a JPEG and thus lose any relevancy that storing the mime type would provide. I will save that issue for another day, if it ever comes up as a significant hindrance to the way images are displayed.

Concerning the unusual URLs, those are, again, a carry-over from the original API, but since I have been wanting to clean the URLs anyhow and your request gave me the effective go-ahead, all "api-manual/solution" paths will now be changed to the more sensible "site/content".

As for some of the more questionable changes you've pointed out, I simply went through and cleaned up any and all references to the old CMS endpoints that I could find. Not only does this clean up the code a little, but it helps ensure that any reliance on the CMS is removed entirely from portal marketplace. If you believe that I am being too forward, I can roll back those changes, though it may have to happen in the next patch, as I am preparing to finalize this next one.

Patch-set: 3
Label: Code-Review=0