Allow for model description to be specified
Covered case where a description can be set when a catalog ID is not
available, resulting in a 400 error due to an undefined value. User will
be unable to use descriptions or documents until a catalog ID has loaded
in or been selected. Also fixes minor issues concerning workflow
display.

Issue-Id: ACUMOS-2839
Change-Id: I6c9592b7cb4b7e6253746ac9600ea4973e16a708
Signed-off-by: swood8 <SW0C74397@TechMahindra.com>