Update patch set 2
Patch Set 2:

ok i will upadte this based on it Tausif.
i will updat the varriables based on the naming convention and based on the condition for undefined and null it is fine as we are already using it for get request i have verified it.

Patch-set: 2
Reviewer: Gerrit User 222 <222@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0