Catch and report catalog name conflicts for peers
When creating a new peer subscription, a local check will now be
performed for the catalog name. If there is no such catalog, the remote
catalog will be mirrored. If the catalog exists and shares the UUID, no
new catalog is made and the process continues as normal. If the catalog
exists but the UUID is different, an error will be thrown.

Issue-Id: ACUMOS-3156
Change-Id: Ie2d29f5db9ea433f43f9ce3a721b783cdcabb125
Signed-off-by: swood8 <SW0C74397@TechMahindra.com>