Update patch set 1
Patch Set 1: Code-Review-1

This does not correctly respond to the original issue. Admins should definitely be able to access on-boarding, but the publisher role is exclusively a supplementary role meant for granting access to the Publish Requests tab. You also perform no check for the "new roles" mentioned in the issue, from which the actual permissions discussed in the issue are meant to come from. Please re-evaluate the issue and try this again.

Patch-set: 1
Reviewer: Gerrit User 101 <101@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1