Update patch set 2
Patch Set 2: Code-Review-1

This is a good start on the core issue and I expect you to add the other permissions to this change such that we can resolve the issue in one Gerrit push, though when I logged in using an admin account that did not have a sub role with "OB" permissions, I found myself unable to access the on-boarding tab. Can you modify this such that admins can access the On-boarding tab as well? If possible, could you also make it to where the tab starts hidden and appears if permissions are available? It might be better to not allow users access to pages they are not allowed to access if their permissions are slow to load in.

Patch-set: 2
Reviewer: Gerrit User 101 <101@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1