Update patch set 2
Patch Set 2:

To compile with Java 11 you must add that version to the Portal job templates in the YAML files controlled in the ci-management repository.  Submit to gerrit and add me as a reviewer.  If you need help on the first one set up a time with me.  Example:

        java-version: openjdk11

Patch-set: 2