Update patch set 2
Patch Set 2:

I recommend you push a new patch to bump the version, since even if you merge, it's unclear to me whether the staging image will be updated. And this would need to be released in a new version anyway. If the new version does not create a snapshot, then you can merge and I will test.

Patch-set: 2
Label: Code-Review=0